+++ Main/Network/Packets/PlayerCommand/CommandAddPacket.cs
@@ -8,7 +8,7 @@ public class CommandAddPacket : BasePacket
 {
     public const PacketId PacketID = PacketId.CommandAdd;
 
-    public int Player;
+    public ushort Player;
     public List<IdentifierData> Units = new();
 
     public override PacketId TypeID => PacketID;
@@ -21,7 +21,7 @@ public class CommandAddPacket : BasePacket
     public override void Send(Buffer writer)
     {
         writer.Write(Player);
-        writer.Write(Units.Count);
+        writer.Write((byte)Units.Count);
         foreach (var unit in Units)
         {
             writer.Write(unit);
@@ -30,8 +30,8 @@ public class CommandAddPacket : BasePacket
 
     public override void Receive(Buffer reader)
     {
-        Player = reader.ReadInt32();
-        var count = reader.ReadInt32();
+        Player = reader.ReadUInt16();
+        var count = reader.ReadByte();
         Units.Clear();
         for (var i = 0; i < count; ++i)
         {
